requirements.txt
GitPython==3.1.40
click==8.1.7
matplotlib==3.8.2
pandas==2.1.4
jinja2==3.1.2